Move time type traits
