+++ crates/chain/chain-type-components/src/traits/mod.rs
@@ -2,3 +2,5 @@ pub mod event;
 pub mod height;
 pub mod ibc;
 pub mod message;
+pub mod time;
+pub mod timeout;
