+++ crates/chain/chain-type-components/src/traits/time.rs
@@ -1,0 +1,6 @@
+use cgp::prelude::*;
+
+#[derive_component(TimeTypeComponent, ProvideTimeType<Chain>)]
+pub trait HasTimeType: Async {
+    type Time: Async;
+}
